Convert homepage to SPA with hash-anchored About + Contact sections
Move About (bio + projects) and Contact (info + form) onto the homepage
as `#about` and `#contact` sections, delete the old `/about` and
`/contact` routes, and add an IntersectionObserver-based scroll spy to
the nav so the active section highlights as the user scrolls. Demos
stays on its own route since it has sub-pages.
