+++ apps/web/components/AboutSection/index.ts
@@ -1,0 +1,1 @@
+export * from './AboutSection';
