+++ apps/web/components/ContactSection/index.ts
@@ -1,0 +1,1 @@
+export * from './ContactSection';
